Finales UI-Design implementiert & diverse andere Änderungen.
Wechsel von float zu double bei Temperatur-Werten. U.a. wegen der neu
verwendeten DoubleBox, welche einen eigenen Parser für Temperaturwerte
unnötig werden lässt. Problem bei DoubleBox und IntegerBox: ',' wird
ignoriert, was verwirrend sein kann (z.B. 1,1 = 11).

Info-Kästchen in der Karte zeigen nun auch genaue Temperaturdifferenz,
wobei die Farbe entsprechend eingefärbt wird.

Tooltips müssen noch so geschrieben werden, dass der User genau weiss,
wozu welcher Wert ist. Insbesondere in der Kartenansicht mit den beiden
Zeitperioden.

+++ ClimateChangeViewer/src/ch/uzh/softwareengineering/climatechangeviewer/client/QueryService.java
@@ -8,10 +8,10 @@ import com.google.gwt.user.client.rpc.RemoteServiceRelativePath;
 public interface QueryService extends RemoteService {
 
 	List<TableDataElement> getTableData(int month, int year1, int year2, String country, String city,
-			float minTemperature, float maxTemperature, float maxTemperatureUncertainty)
+			double minTemperature, double maxTemperature, double maxTemperatureUncertainty)
 					throws FilterOverflowException, NoEntriesFoundException, DataFileCorruptedException;
 	
-	List<MapDataElement> getMapData(int comparisonPerdiod1Start , int comparisonPerdiod2Start, float maxTemperatureUncertainty)
+	List<MapDataElement> getMapData(int comparisonPerdiod1Start , int comparisonPerdiod2Start, double maxTemperatureUncertainty)
 			throws NoEntriesFoundException, DataFileCorruptedException;
 	
 }
