+++ ClimateChangeViewer/src/ch/uzh/softwareengineering/climatechangeviewer/server/CoordinatesParser.java
@@ -2,31 +2,31 @@ package ch.uzh.softwareengineering.climatechangeviewer.server;
 
 public class CoordinatesParser {
 	
-	public static float parseLatitude(String latitudeString) {
+	public static double parseLatitude(String latitudeString) {
 		String latitudeDigitsString;
-		float latitude = Float.MAX_VALUE;
+		double latitude = Double.MAX_VALUE;
 		
 		if(latitudeString.toUpperCase().indexOf('N') >= 0) {
 			latitudeDigitsString = latitudeString.substring(0, latitudeString.toUpperCase().indexOf('N'));
-			latitude = Float.parseFloat(latitudeDigitsString);
+			latitude = Double.parseDouble(latitudeDigitsString);
 		} else if(latitudeString.toUpperCase().indexOf('S') >= 0)  {
 			latitudeDigitsString = latitudeString.substring(0, latitudeString.toUpperCase().indexOf('S'));
-			latitude = Float.parseFloat(latitudeDigitsString);
+			latitude = Double.parseDouble(latitudeDigitsString);
 			latitude = latitude * (-1);
 		}
 		return latitude;
 	}
 	
-	public static float parseLongitude(String longitudeString) {
+	public static double parseLongitude(String longitudeString) {
 		String longitudeDigitsString;
-		float longitude = Float.MAX_VALUE;
+		double longitude = Double.MAX_VALUE;
 		
 		if(longitudeString.toUpperCase().indexOf('E') >= 0) {
 			longitudeDigitsString = longitudeString.substring(0, longitudeString.toUpperCase().indexOf('E'));
-			longitude = Float.parseFloat(longitudeDigitsString);
+			longitude = Double.parseDouble(longitudeDigitsString);
 		} else if (longitudeString.toUpperCase().indexOf('W') >= 0) {
 			longitudeDigitsString = longitudeString.substring(0, longitudeString.toUpperCase().indexOf('W'));
-			longitude = Float.parseFloat(longitudeDigitsString);
+			longitude = Double.parseDouble(longitudeDigitsString);
 			longitude = longitude * (-1);
 		}
 		return longitude;
