+++ ClimateChangeViewer/src/ch/uzh/softwareengineering/climatechangeviewer/server/DataFileLine.java
@@ -8,10 +8,10 @@ public class DataFileLine {
 	private String country = "";
 	private int month = Integer.MIN_VALUE;
 	private int year = Integer.MIN_VALUE;
-	private float uncertainty = Float.MAX_VALUE;
-	private float temperature = Float.MAX_VALUE;
-	private float longitude = Float.MAX_VALUE;
-	private float latitude = Float.MAX_VALUE;
+	private double uncertainty = Double.MAX_VALUE;
+	private double temperature = Double.MAX_VALUE;
+	private double longitude = Double.MAX_VALUE;
+	private double latitude = Double.MAX_VALUE;
 	
 	public int getMonth() {
 		return month;
@@ -45,35 +45,35 @@ public class DataFileLine {
 		this.country = country;
 	}
 	
-	public float getUncertainty() {
+	public double getUncertainty() {
 		return uncertainty;
 	}
 	
-	public void setUncertainty(float uncertainty) {
+	public void setUncertainty(double uncertainty) {
 		this.uncertainty = uncertainty;
 	}
 	
-	public float getTemperature() {
+	public double getTemperature() {
 		return temperature;
 	}
 	
-	public void setTemperature(float temperature) {
+	public void setTemperature(double temperature) {
 		this.temperature = temperature;
 	}
 	
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 	
-	public void setLongitude(float langitude) {
+	public void setLongitude(double langitude) {
 		this.longitude = langitude;
 	}
 	
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 	
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
