+++ ClimateChangeViewer/test/ch/uzh/softwareengineering/climatechangeviewer/client/TableDataElementTestCase.java
@@ -55,15 +55,15 @@ public class TableDataElementTestCase extends GWTTestCase {
 	public void testGetTemperatureString() {
 		TableDataElement dataElement = new TableDataElement();
 		
-		float temperature1 = 12.123f;
+		double temperature1 = 12.123f;
 		dataElement.setTemperature(temperature1);
 		assertTrue(dataElement.getTemperatureString().equals("12.123"));
 		
-		float temperature2 = 1f;
+		double temperature2 = 1f;
 		dataElement.setTemperature(temperature2);
 		assertTrue(dataElement.getTemperatureString().equals("1.000"));
 		
-		float temperature3 = -0.001f;
+		double temperature3 = -0.001f;
 		dataElement.setTemperature(temperature3);
 		assertTrue(dataElement.getTemperatureString().equals("-0.001"));
 	}
@@ -72,16 +72,16 @@ public class TableDataElementTestCase extends GWTTestCase {
 	public void testGetTemperatureUncertaintyString() {
 		TableDataElement dataElement = new TableDataElement();
 		
-		float temperatureUncertainty1 = 12.123f;
+		double temperatureUncertainty1 = 12.123f;
 		dataElement.setUncertainty(temperatureUncertainty1);
 		System.out.println(dataElement.getUncertaintyString());
 		assertTrue(dataElement.getUncertaintyString().equals("12.123"));
 		
-		float temperatureUncertainty2 = 1f;
+		double temperatureUncertainty2 = 1f;
 		dataElement.setUncertainty(temperatureUncertainty2);
 		assertTrue(dataElement.getUncertaintyString().equals("1.000"));
 		
-		float temperatureUncertainty3 = 0.01f;
+		double temperatureUncertainty3 = 0.01f;
 		dataElement.setUncertainty(temperatureUncertainty3);
 		assertTrue(dataElement.getUncertaintyString().equals("0.010"));
 	}
