+++ ClimateChangeViewer/test/ch/uzh/softwareengineering/climatechangeviewer/server/CoordinatesParserTestCase.java
@@ -9,7 +9,7 @@ public class CoordinatesParserTestCase {
 	@Test
 	public void testParseLatitude() {
 		// The method will fail if the string parameter has another format than the used data file. For empty
-		// strings Float.MAX_VALUE is assigned. 
+		// strings Double.MAX_VALUE is assigned. 
 		String latitudeString1 = "";
 		String latitudeString2 = " ";
 		String latitudeString3 = "0";
@@ -22,9 +22,9 @@ public class CoordinatesParserTestCase {
 		String latitudeString10 = ".403S";
 		String latitudeString11 = ".403N";
 
-		assertEquals(Float.MAX_VALUE, CoordinatesParser.parseLatitude(latitudeString1), 0.001f);
-		assertEquals(Float.MAX_VALUE, CoordinatesParser.parseLatitude(latitudeString2), 0.001f);
-		assertEquals(Float.MAX_VALUE, CoordinatesParser.parseLatitude(latitudeString3), 0.001f);
+		assertEquals(Double.MAX_VALUE, CoordinatesParser.parseLatitude(latitudeString1), 0.001f);
+		assertEquals(Double.MAX_VALUE, CoordinatesParser.parseLatitude(latitudeString2), 0.001f);
+		assertEquals(Double.MAX_VALUE, CoordinatesParser.parseLatitude(latitudeString3), 0.001f);
 		assertEquals(0f, CoordinatesParser.parseLatitude(latitudeString4), 0.001f);
 		assertEquals(0f, CoordinatesParser.parseLatitude(latitudeString5), 0.001f);
 		assertEquals(-124.403, CoordinatesParser.parseLatitude(latitudeString6), 0.001f);
@@ -38,7 +38,7 @@ public class CoordinatesParserTestCase {
 	@Test
 	public void testParseLongitude() {
 		// The method will fail if the string parameter has another format than the used data file. For empty
-		// strings Float.MAX_VALUE is assigned. 
+		// strings Double.MAX_VALUE is assigned. 
 		String longitudeString1 = "";
 		String longitudeString2 = " ";
 		String longitudeString3 = "0";
@@ -51,9 +51,9 @@ public class CoordinatesParserTestCase {
 		String longitudeString10 = ".403W";
 		String longitudeString11 = ".403E";
 
-		assertEquals(Float.MAX_VALUE, CoordinatesParser.parseLongitude(longitudeString1), 0.001f);
-		assertEquals(Float.MAX_VALUE, CoordinatesParser.parseLongitude(longitudeString2), 0.001f);
-		assertEquals(Float.MAX_VALUE, CoordinatesParser.parseLongitude(longitudeString3), 0.001f);
+		assertEquals(Double.MAX_VALUE, CoordinatesParser.parseLongitude(longitudeString1), 0.001f);
+		assertEquals(Double.MAX_VALUE, CoordinatesParser.parseLongitude(longitudeString2), 0.001f);
+		assertEquals(Double.MAX_VALUE, CoordinatesParser.parseLongitude(longitudeString3), 0.001f);
 		assertEquals(0f, CoordinatesParser.parseLongitude(longitudeString4), 0.001f);
 		assertEquals(0f, CoordinatesParser.parseLongitude(longitudeString5), 0.001f);
 		assertEquals(-124.403, CoordinatesParser.parseLongitude(longitudeString6), 0.001f);
